requirements.txt
mss==9.0.1
pyautogui==0.9.54
pynput==1.7.6
fpdf2==2.7.6
PySide6==6.8.1.1
pyinstaller==6.11.0 